# openapi.yaml
openapi: 3.0.0
info:
  title: vm-runner
  version: 0.1.1
servers: []
tags: []
paths:
  /healthz:
    get:
      responses:
        '200':
          description: ''
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
        '503':
          description: ''
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
  /v1/sessions:
    post:
      requestBody:
        content:
          application/json; charset=utf-8:
            schema:
              $ref: '#/components/schemas/Create'
        required: true
      responses:
        '201':
          description: ''
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/SessionBody'
        '200':
          description: ''
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/SessionBody'
        '409':
          description: ''
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/Conflict'
        '400':
          description: ''
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/NotExecuted'
        '401':
          description: ''
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/Refusal'
        '503':
          description: ''
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
  /v1/whoami:
    get:
      responses:
        '200':
          description: ''
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/Identity'
        '401':
          description: ''
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/Refusal'
  /v1/sessions/{id}/exec:
    post:
      parameters:
      - name: id
        schema:
          type: string
        in: path
        required: true
        deprecated: false
        explode: true
      requestBody:
        content:
          application/json; charset=utf-8:
            schema:
              $ref: '#/components/schemas/Exec'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/ExecResult'
        '202':
          description: ''
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/Pending'
        '401':
          description: ''
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/Refusal'
        '400':
          description: ''
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/ExecResult'
        '413':
          description: ''
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/ExecResult'
        '404':
          description: ''
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
        '503':
          description: ''
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
        '502':
          description: ''
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/ExecResult'
  /v1/jobs/{jobId}:
    get:
      parameters:
      - name: jobId
        schema:
          type: string
        in: path
        required: true
        deprecated: false
        explode: true
      responses:
        '200':
          description: ''
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/JobStatus'
        '401':
          description: ''
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/Refusal'
        '404':
          description: ''
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
        '503':
          description: ''
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
        '502':
          description: ''
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/ExecResult'
  /exec:
    post:
      requestBody:
        content:
          application/json; charset=utf-8:
            schema:
              $ref: '#/components/schemas/ExecInput'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/Terminal'
        '202':
          description: ''
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/Unknown'
  /jobs/{id}:
    get:
      parameters:
      - name: id
        schema:
          type: string
        in: path
        required: true
        deprecated: false
        explode: true
      responses:
        '200':
          description: ''
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/View'
        '202':
          description: ''
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/Unknown'
components:
  schemas:
    Conflict:
      type: object
      title: Conflict
      required:
      - error
      properties:
        error:
          $ref: '#/components/schemas/ConflictCode'
    ConflictCode:
      type: string
      enum:
      - session_profile_conflict
    ControllerExecuted:
      type: object
      title: ControllerExecuted
      required:
      - exitCode
      - stdout
      - stderr
      - truncated
      properties:
        exitCode:
          type: integer
          format: int32
        stdout:
          type: string
        stderr:
          type: string
        truncated:
          type: boolean
    ControllerRunning:
      type: object
      title: ControllerRunning
      required:
      - state
      properties:
        state:
          $ref: '#/components/schemas/RunningState'
    ControllerUnknown:
      type: string
      enum:
      - unknown
    Create:
      type: object
      title: Create
      required:
      - profile
      properties:
        profile:
          type: string
        name:
          type: string
    Exec:
      type: object
      title: Exec
      required:
      - argv
      - deadlineMs
      properties:
        argv:
          type: array
          items:
            type: string
        stdin:
          type: string
        deadlineMs:
          type: integer
          format: uint32
          maximum: 25000.0
          minimum: 1000.0
    ExecInput:
      type: object
      title: ExecInput
      required:
      - argv
      - deadlineMs
      properties:
        argv:
          type: array
          items:
            type: string
        stdin:
          type: string
        deadlineMs:
          type: integer
          format: uint32
          maximum: 25000.0
    ExecRefused:
      type: object
      title: ExecRefused
      required:
      - reason
      - truncated
      properties:
        reason:
          type: string
        truncated:
          type: boolean
    ExecResult:
      type: object
      anyOf:
      - $ref: '#/components/schemas/ExecResult_ControllerExecuted'
      - $ref: '#/components/schemas/ExecResult_ExecRefused'
      discriminator:
        propertyName: outcome
        mapping:
          executed: '#/components/schemas/ExecResult_ControllerExecuted'
          not_executed: '#/components/schemas/ExecResult_ExecRefused'
    ExecResult_ControllerExecuted:
      allOf:
      - type: object
        required:
        - outcome
        properties:
          outcome:
            type: string
            enum:
            - executed
            example: executed
      - $ref: '#/components/schemas/ControllerExecuted'
    ExecResult_ExecRefused:
      allOf:
      - type: object
        required:
        - outcome
        properties:
          outcome:
            type: string
            enum:
            - not_executed
            example: not_executed
      - $ref: '#/components/schemas/ExecRefused'
    Executed:
      type: object
      title: Executed
      required:
      - exitCode
      - stdout
      - stderr
      - truncated
      properties:
        exitCode:
          type: integer
          format: int32
        stdout:
          type: string
        stderr:
          type: string
        truncated:
          type: boolean
    Failure:
      type: string
      enum:
      - quota
      - bad_profile
      - bad_name
    Identity:
      type: object
      title: Identity
      required:
      - subject
      - quota
      properties:
        subject:
          type: string
        quota:
          $ref: '#/components/schemas/Quota'
    JailNotExecuted:
      type: object
      title: JailNotExecuted
      required:
      - reason
      - truncated
      properties:
        reason:
          type: string
        truncated:
          type: boolean
    JobStatus:
      type: object
      anyOf:
      - $ref: '#/components/schemas/ControllerRunning'
      - $ref: '#/components/schemas/ExecResult'
    NotExecuted:
      type: object
      title: NotExecuted
      required:
      - outcome
      - reason
      properties:
        outcome:
          $ref: '#/components/schemas/NotExecutedOutcome'
        reason:
          $ref: '#/components/schemas/Failure'
    NotExecutedOutcome:
      type: string
      enum:
      - not_executed
    Pending:
      type: object
      title: Pending
      required:
      - outcome
      - jobId
      properties:
        outcome:
          $ref: '#/components/schemas/ControllerUnknown'
        jobId:
          type: string
    Quota:
      type: object
      title: Quota
      required:
      - maxSessions
      properties:
        maxSessions:
          type: integer
          format: uint32
    Reason:
      type: string
      enum:
      - audience
      - issuer
      - expired
      - not_yet_valid
      - signature
      - unknown_subject
      - malformed
    Refusal:
      type: object
      title: Refusal
      required:
      - error
      - reason
      properties:
        error:
          type: string
        reason:
          $ref: '#/components/schemas/Reason'
    Running:
      type: object
      title: Running
      required:
      - state
      properties:
        state:
          type: string
    RunningState:
      type: string
      enum:
      - running
    SessionBody:
      type: object
      title: SessionBody
      required:
      - sessionId
      - name
      - profile
      - shape
      - state
      properties:
        sessionId:
          type: string
        name:
          type: string
        profile:
          type: string
        shape:
          type: string
        state:
          $ref: '#/components/schemas/SessionState'
    SessionState:
      type: string
      enum:
      - pending
      - ready
    Terminal:
      type: object
      anyOf:
      - $ref: '#/components/schemas/Terminal_Executed'
      - $ref: '#/components/schemas/Terminal_JailNotExecuted'
      discriminator:
        propertyName: outcome
        mapping:
          executed: '#/components/schemas/Terminal_Executed'
          not_executed: '#/components/schemas/Terminal_JailNotExecuted'
    Terminal_Executed:
      allOf:
      - type: object
        required:
        - outcome
        properties:
          outcome:
            type: string
            enum:
            - executed
            example: executed
      - $ref: '#/components/schemas/Executed'
    Terminal_JailNotExecuted:
      allOf:
      - type: object
        required:
        - outcome
        properties:
          outcome:
            type: string
            enum:
            - not_executed
            example: not_executed
      - $ref: '#/components/schemas/JailNotExecuted'
    Unknown:
      type: object
      title: Unknown
      required:
      - outcome
      - jobId
      properties:
        outcome:
          type: string
        jobId:
          type: string
    View:
      type: object
      anyOf:
      - $ref: '#/components/schemas/Running'
      - $ref: '#/components/schemas/Terminal'